gnhf 35: Relocated the dead-within-performance.py _ALIAS map and _num parser to ingestion/load_cdp_cost.py, their only consumer, removing an inverted cross-layer import, behavior-preserving.
